Add 'sports' provider module
