+++ provider/sports/src/main/kotlin/io/github/serpro69/kfaker/sports/provider/Crossfit.kt
@@ -1,13 +1,16 @@
-package io.github.serpro69.kfaker.provider
+package io.github.serpro69.kfaker.sports.provider
 
 import io.github.serpro69.kfaker.FakerService
 import io.github.serpro69.kfaker.dictionary.YamlCategory
+import io.github.serpro69.kfaker.provider.FakeDataProvider
+import io.github.serpro69.kfaker.provider.YamlFakeDataProvider
 import io.github.serpro69.kfaker.provider.unique.LocalUniqueDataProvider
 import io.github.serpro69.kfaker.provider.unique.UniqueProviderDelegate
 
 /**
  * [FakeDataProvider] implementation for [YamlCategory.CROSSFIT] category.
  */
+@Suppress("unused")
 class Crossfit internal constructor(fakerService: FakerService): YamlFakeDataProvider<Crossfit>(fakerService){
     override val yamlCategory = YamlCategory.CROSSFIT
     override val localUniqueDataProvider= LocalUniqueDataProvider<Crossfit>()
